metacognition-agent/requirements.txt
# ADK Framework
google-adk>=0.1.0

# A2A Protocol SDK
a2a-sdk>=0.2.0

# Git operations
GitPython>=3.1.0

# Async operations and HTTP
aiohttp>=3.8.0
asyncio
python-dotenv>=0.19.0

# Data handling and utilities
pathlib2>=2.3.0
PyYAML>=6.0
structlog>=21.0.0

# Optional: For enhanced functionality
prometheus-client>=0.12.0 